test: fold the rest onto the option, and drop the last whole-file skip
The image minimizers, the query in an asset name and what identifies a
minimizer to the cache are all the `minify` option: which assets it is
offered and what it does with them. They join it, and the worker joins
`parallel`, which is the option that decides whether there is one.

That leaves one whole-file skip instead of two. The image file was skipped
wholesale where `sharp` and friends are absent; nothing in it reaches them
at module scope, so it is gated on its block instead and every other case in
`minify-option` now runs on the legacy rows too. Only `embedded-source` is
still skipped whole, because it requires a deep webpack path at module scope
that ships from 5.110.

Thirteen files where there were twenty-four, same 573 tests and 825
snapshots. `minify-option` carries 157 of them, which is the cost of the
option being this large a subject.
