+++ jest.config.js
@@ -1,21 +1,14 @@
 // Which suites this environment can run is decided in `test/helpers/env.js`, so
 // the same answer gates a whole file here and a single block inside a file.
-const {
-  RUN_EMBEDDED_TESTS,
-  RUN_IMAGE_TESTS,
-} = require("./test/helpers/env.js");
+const { RUN_EMBEDDED_TESTS } = require("./test/helpers/env.js");
 
 const testPathIgnorePatterns = [];
 
-// Both of these reach for what they need at module scope — a deep webpack path
-// that only ships from 5.110, and packages the legacy rows never install — so
-// they are skipped whole rather than block by block.
+// The one file that reaches for what it needs at module scope — a deep webpack
+// path that only ships from 5.110 — so it is skipped whole rather than block by
+// block. Everywhere else the gate is on the block.
 if (!RUN_EMBEDDED_TESTS) {
   testPathIgnorePatterns.push("/test/embedded-source\\.test\\.js$");
-}
-
-if (!RUN_IMAGE_TESTS) {
-  testPathIgnorePatterns.push("/test/image-minify-option\\.test\\.js$");
 }
 
 module.exports = {
